Crouch movement, Animations system improvements
Added:
- Crouch movement system with Animations, state, etc
- Slightly refactored movement system (Walk, Sprint logic)

+++ Assets/!Project/Scripts/Data/PlayerStatsConfig.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class PlayerStatsConfig: SerializedScriptableObject
 {
     public int Health = 0;
-    public int MoveSpeed = 0;
+    public float MoveSpeed = 0;
     public int JumpPower = 0;
 
     public PlayerStatsConfig Clone()
